backend/requirements.txt
# Core API
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic[email]>=2.4.2
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0
cryptography>=41.0.0

# Networking & Scanning
httpx>=0.25.1
requests>=2.31.0
dnspython>=2.4.2
python-magic>=0.4.27

# OSINT & Tools
holehe>=1.61.1
sherlock-project>=0.14.3
psutil>=5.9.6

# Utilities
jinja2>=3.1.2
colorama>=0.4.6

python-whois>=0.9.34
uuid>=1.30
